+++ scripts/oauth/oauth.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -e
 
 # --- Functions
 
@@ -31,6 +32,7 @@ assert_users_and_oauth() {
     USERS=$(psql -t -c "SELECT COUNT(*) FROM user_oauth;")
     OAUTH_ACCOUNTS=$(psql -t -c "SELECT COUNT(*) FROM oauthaccount;")
     if [ "$USERS" -ne "$1" ] || [ "$OAUTH_ACCOUNTS" -ne "$2" ]; then
+        echo "ERROR: Expected (users, oauth_accounts)==(${1}, ${2}), got (${USERS}, ${OAUTH_ACCOUNTS})."
         exit 1
     fi
 }
@@ -46,21 +48,33 @@ assert_email_and_id(){
     EMAIL=$(echo $USER | jq -r ".email")
     ID=$(echo $USER | jq -r ".id")
     if [ "$EMAIL" != "$2" ]; then
+        echo "ERROR: Expected email==${2}, got ${EMAIL}."
         exit 1
     fi
     if [ "$ID" != "$3" ]; then
+        echo "ERROR: Expected user_id==${3}, got ${ID}."
+        exit 1
+    fi
+}
+
+assert_email_count(){
+    NUM_MESSAGE=$(
+        curl --silent http://localhost:8025/api/v1/messages | jq -r ".total"
+    )
+    if [ "$NUM_MESSAGE" -ne "$1" ]; then
+        echo "ERROR: Expected email_count==${1}, got ${NUM_MESSAGE}."
         exit 1
     fi
 }
 
 # --- Test
-
+assert_users_and_oauth 1 0
+assert_email_count 0
 
 # Register "kilgore@kilgore.trout" (the user from Dex) as regular account.
 SUPERUSER_TOKEN=$(standard_login "admin@fractal.xy" "1234")
 
-assert_users_and_oauth 1 0
-curl -X POST \
+curl --silent -X POST \
     http://127.0.0.1:8001/auth/register/ \
     -H "Content-Type: application/json" \
     -H "Authorization: Bearer $SUPERUSER_TOKEN" \
@@ -85,7 +99,7 @@ assert_users_and_oauth 2 1
 assert_email_and_id $USER_TOKEN_OAUTH "kilgore@kilgore.trout" $USER_ID
 
 # Change email into "kilgore@fractal.xy".
-curl -X PATCH \
+curl --silent -X PATCH \
     "http://127.0.0.1:8001/auth/users/$USER_ID/" \
     -H "Content-Type: application/json" \
     -H "Authorization: Bearer $SUPERUSER_TOKEN" \
@@ -109,7 +123,12 @@ assert_email_and_id $USER_TOKEN "kilgore@fractal.xy" $USER_ID
 
 # Using oauth login creates another user: "kilgore@kilgore.trout".
 assert_users_and_oauth 2 0
+assert_email_count 0
 USER_TOKEN_OAUTH=$(oauth_login)
 assert_users_and_oauth 3 1
+assert_email_count 1
+# Print emails
+echo EMAIL LIST
+curl --silent http://localhost:8025/api/v1/messages | jq
 
 assert_email_and_id $USER_TOKEN_OAUTH "kilgore@kilgore.trout" $((USER_ID+1))